Cleanup vue
