+++ npm/vue/src/index.ts
@@ -6,6 +6,7 @@ import {
   VueTestUtilsConfigOptions,
   Wrapper,
 } from '@vue/test-utils'
+import { renderTestingPlatform } from './renderTestingPlatform'
 
 const defaultOptions: (keyof MountOptions)[] = [
   'vue',
@@ -361,11 +362,7 @@ export const mount = (
 
     // If the target div doesn't exist, create it
     if (!el) {
-      const div = document.createElement('div')
-
-      div.id = 'cypress-jsdom'
-      document.body.appendChild(div)
-      el = div
+      renderTestingPlatform(document.head.innerHTML)
     }
 
     if (typeof options.stylesheets === 'string') {
